requirements.txt
sty==1.0.0rc1
pycryptodome==3.10.1

